add test for smallest range

+++ sources/smallest_range/smallest_range_solution2.cpp
@@ -3,7 +3,7 @@ int smallest_range_II(std::vector<int>& A, const int K)
 {
   std::sort(std::begin(A), std::end(A));
   int ans = A.back() - A.front();
-  for (int i = 1; i < A.size(); i++)
+  for (size_t i = 1; i < A.size(); i++)
   {
     const auto m_new = std::min(A.front() + K, A[i] - K);
     const auto M_new = std::max(A[i - 1] + K, A.back() - K);
